Move away from shippable
They changed their API, and I don't really care to update the code to
work with it. For now we will just use Gitlab CI and just have the Gitlab
mirror update itself and run the CI pipelines in its own time. It won't
run immediately, but that isn't the end of the world.